Prepare unit test

+++ controllers/nodehealthcheck_controller_test.go
@@ -181,11 +181,13 @@ var _ = Describe("Node Health Check CR", func() {
 			deleteObjects(objects...)
 
 			// delete all remediation CRs
-			cr := newRemediationCR("")
-			crList := &unstructured.UnstructuredList{Object: cr.Object}
-			Expect(k8sClient.List(context.Background(), crList)).To(Succeed())
-			for _, item := range crList.Items {
-				Expect(k8sClient.Delete(context.Background(), &item)).To(Succeed())
+			if underTest.Spec.RemediationTemplate.Kind != "dummyTemplate" {
+				cr := newRemediationCR("", underTest)
+				crList := &unstructured.UnstructuredList{Object: cr.Object}
+				Expect(k8sClient.List(context.Background(), crList)).To(Succeed())
+				for _, item := range crList.Items {
+					Expect(k8sClient.Delete(context.Background(), &item)).To(Succeed())
+				}
 			}
 
 			// let thing settle a bit
@@ -198,7 +200,7 @@ var _ = Describe("Node Health Check CR", func() {
 			})
 
 			It("create a remediation CR for each unhealthy node and updates status", func() {
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(cr.Object).To(ContainElement(map[string]interface{}{"size": "foo"}))
@@ -208,6 +210,7 @@ var _ = Describe("Node Health Check CR", func() {
 							// Kind and API version aren't set on underTest, envtest issue...
 							// Controller is empty for HaveField because false is the zero value?
 							HaveField("Name", underTest.Name),
+							HaveField("UID", underTest.UID),
 						),
 					))
 				Expect(cr.GetAnnotations()[oldRemediationCRAnnotationKey]).To(BeEmpty())
@@ -234,7 +237,7 @@ var _ = Describe("Node Health Check CR", func() {
 			})
 
 			It("skips remediation - CR is not created, status updated correctly", func() {
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(errors.IsNotFound(err)).To(BeTrue())
 
@@ -250,8 +253,8 @@ var _ = Describe("Node Health Check CR", func() {
 		When("few nodes become healthy", func() {
 			BeforeEach(func() {
 				setupObjects(1, 2)
-				remediationCR := newRemediationCR("healthy-worker-node-2")
-				remediationCROther := newRemediationCR("healthy-worker-node-1")
+				remediationCR := newRemediationCR("healthy-worker-node-2", underTest)
+				remediationCROther := newRemediationCR("healthy-worker-node-1", underTest)
 				refs := remediationCROther.GetOwnerReferences()
 				refs[0].Name = "other"
 				remediationCROther.SetOwnerReferences(refs)
@@ -259,16 +262,16 @@ var _ = Describe("Node Health Check CR", func() {
 			})
 
 			It("deletes an existing remediation CR and updates status", func() {
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(err).NotTo(HaveOccurred())
 
-				cr = newRemediationCR("healthy-worker-node-2")
+				cr = newRemediationCR("healthy-worker-node-2", underTest)
 				err = k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(errors.IsNotFound(err)).To(BeTrue())
 
 				// owned by other NHC, should not be deleted
-				cr = newRemediationCR("healthy-worker-node-1")
+				cr = newRemediationCR("healthy-worker-node-1", underTest)
 				err = k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(err).NotTo(HaveOccurred())
 
@@ -302,7 +305,7 @@ var _ = Describe("Node Health Check CR", func() {
 				Expect(k8sClient.Update(context.Background(), underTest)).To(Succeed())
 				time.Sleep(2 * time.Second)
 
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(cr.GetAnnotations()[oldRemediationCRAnnotationKey]).To(Equal("flagon"))
@@ -319,7 +322,7 @@ var _ = Describe("Node Health Check CR", func() {
 				})
 
 				It("creates a one remediation CR for control plane node and updates status", func() {
-					cr := newRemediationCR("")
+					cr := newRemediationCR("", underTest)
 					crList := &unstructured.UnstructuredList{Object: cr.Object}
 					Expect(k8sClient.List(context.Background(), crList)).To(Succeed())
 
@@ -344,7 +347,7 @@ var _ = Describe("Node Health Check CR", func() {
 			})
 
 			It("skips remediation and updates status", func() {
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(errors.IsNotFound(err)).To(BeTrue())
 
@@ -368,7 +371,7 @@ var _ = Describe("Node Health Check CR", func() {
 			})
 
 			It("doesn't not remediate but requeues reconciliation and updates status", func() {
-				cr := newRemediationCR("unhealthy-worker-node-1")
+				cr := newRemediationCR("unhealthy-worker-node-1", underTest)
 				err := k8sClient.Get(context.Background(), client.ObjectKeyFromObject(cr), cr)
 				Expect(errors.IsNotFound(err)).To(BeTrue())
 
@@ -395,7 +398,7 @@ var _ = Describe("Node Health Check CR", func() {
 		When("Nodes are candidates for remediation but remediation template is broken", func() {
 			BeforeEach(func() {
 				setupObjects(1, 2)
-				underTest.Spec.RemediationTemplate.Kind = "dummy"
+				underTest.Spec.RemediationTemplate.Kind = "dummyTemplate"
 			})
 
 			It("should set corresponding condition", func() {
@@ -492,25 +495,28 @@ var _ = Describe("Node Health Check CR", func() {
 	})
 })
 
-func newRemediationCR(nodeName string) *unstructured.Unstructured {
-	return newRemediationCRWithRole(nodeName, false)
+func newRemediationCR(nodeName string, nhc *v1alpha1.NodeHealthCheck) *unstructured.Unstructured {
+	return newRemediationCRWithRole(nodeName, nhc, false)
 }
 
-func newRemediationCRWithRole(nodeName string, isControlPlaneNode bool) *unstructured.Unstructured {
+func newRemediationCRWithRole(nodeName string, nhc *v1alpha1.NodeHealthCheck, isControlPlaneNode bool) *unstructured.Unstructured {
 	cr := unstructured.Unstructured{}
 	cr.SetName(nodeName)
-	cr.SetNamespace("default")
+	cr.SetNamespace(nhc.Spec.RemediationTemplate.Namespace)
+	kind := nhc.Spec.RemediationTemplate.GroupVersionKind().Kind
+	// remove trailing template
+	kind = kind[:len(kind)-len("template")]
 	cr.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   testRemediationCRD.Spec.Group,
-		Version: testRemediationCRD.Spec.Versions[0].Name,
-		Kind:    testRemediationCRD.Spec.Names.Kind,
+		Group:   nhc.Spec.RemediationTemplate.GroupVersionKind().Group,
+		Version: nhc.Spec.RemediationTemplate.GroupVersionKind().Version,
+		Kind:    kind,
 	})
 	cr.SetOwnerReferences([]metav1.OwnerReference{
 		{
-			APIVersion: "remediation.medik8s.io/v1alpha1",
-			Kind:       "NodeHealthCheck",
-			Name:       "test",
-			UID:        "1234",
+			APIVersion: nhc.APIVersion,
+			Kind:       nhc.Kind,
+			Name:       nhc.Name,
+			UID:        nhc.UID,
 		},
 	})
 	if isControlPlaneNode {
